LPOPDY: BETIN now bit-exact (Fortran single-prec Stirling constant) + op-order fixes
Root-caused the BETIN 1.56e-8 residual: mpbgam.f's Stirling constant 0.9189385332046727 has NO
'D' suffix -> gfortran reads it as a SINGLE-precision literal (0.91893852), not double. Matching
that (Float64(0.9189385332046727f0)) makes the whole BETIN package 30/30 BIT-EXACT (was 1.56e-8).
Also tightened mpb_mpbmod op-order to gfortran: PIODEN=B1INC/E1 computed in Float32 then widened;
AGG=TREES*BETIN computed in DOUBLE then rounded to REAL. Chain golden test 45/45, BETIN tol now ==0.

+++ scratchpad/lpmpb/betin_pkg.jl
@@ -17,7 +17,7 @@ function mpbgam(xx::Float64)::Float64
             term *= zz; zz += 1.0
         end
         rz2 = 1.0/zz^2
-        dlgam = (zz-0.5)*log(zz) - zz + 0.9189385332046727 - log(term) +
+        dlgam = (zz-0.5)*log(zz) - zz + Float64(0.9189385332046727f0) - log(term) +
             (1.0/zz)*(0.8333333333333333e-1 - (rz2*(0.2777777777777777e-2 +
             (rz2*(0.7936507936507936e-3 - (rz2*(0.5952380952380952e-3)))))))
     end
